fix(rgcp/lodeom): corrige la saisie de décimales en mois par mois
